# api-specs/openapi-specs/OpenApi.yml
openapi: 3.0.3
info:
  title: Course Management API
  version: 1.0.0
  description: >
    This API allows users to authenticate using Keycloak and perform operations
    related to course creation, review submission, and fetching course details with reviews.
  contact: {}
servers:
  - url: http://localhost:8081
  - url: http://localhost:9000
paths:
  /realms/course-management-realm/protocol/openid-connect/token:
    post:
      summary: Obtain Access Token
      description: >
        This endpoint allows users to authenticate via Keycloak using Resource Owner Password Credentials Grant
        and obtain a JWT access token.
      operationId: getAccessToken
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                client_id:
                  type: string
                  example: course-app
                client_secret:
                  type: string
                  example: v1sCIPjANbvyJ87RsTkYeI9xHonDqZh7
                grant_type:
                  type: string
                  example: password
                password:
                  type: string
                  example: password
                scope:
                  type: string
                  example: openid roles
                username:
                  type: string
                  example: kapil
            examples:
              AccessToken:
                value:
                  client_id: course-app
                  client_secret: v1sCIPjANbvyJ87RsTkYeI9xHonDqZh7
                  grant_type: password
                  password: password
                  scope: openid roles
                  username: kapil
      responses:
        '200':
          description: Successfully retrieved access token
      security:
        - {}
  /courses:
    post:
      summary: Create a New Course
      description: >
          This endpoint allows authenticated users to create a new course by providing the title, author, publisher, and price.
      operationId: addCourse
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                author:
                  type: string
                  example: John Doe
                price:
                  type: number
                  example: 29.79
                publisher:
                  type: string
                  example: Whatsapp
                title:
                  type: string
                  example: Microservices with Quarkus
            examples:
              Add Course:
                value:
                  author: John Doe
                  price: 29.79
                  publisher: Whatsapp
                  title: Microservices with Quarkus
      responses:
        '200':
          description: Course created successfully
      security:
        - bearerAuth: []
  /reviews:
    post:
      summary: Submit a Review for a Course
      description: >
          This endpoint allows authenticated users to submit a review for an existing course by specifying author, content, courseId, and email.
      operationId: addReview
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                author:
                  type: string
                  example: John Doe
                content:
                  type: string
                  example: Amazing book and loved reading it
                courseId:
                  type: number
                  example: 2
                email:
                  type: string
                  example: abc@xyz.com
            examples:
              Add Review:
                value:
                  author: John Doe
                  content: Amazing book and loved reading it
                  courseId: 2
                  email: abc@xyz.com
      responses:
        '200':
          description: Review added successfully
      security:
        - bearerAuth: []
  /course-aggregate/4/with-details:
    get:
      summary: Get Course Details with Reviews
      description: >
        This endpoint retrieves a course's full details along with all associated user reviews.
      operationId: getCourseComposite
      parameters:
        - name: ''
          in: header
          schema:
            type: string
            example: ''
      responses:
        '200':
          description: Successfully retrieved course and associated reviews
      security:
        - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
tags: []